+++ src/components/Insights/tickets/common/NPlusOneAffectedEndpoints/types.ts
@@ -1,5 +1,5 @@
 import { SpanNPlusOneInsight } from "../../../types";
 
 export interface NPlusOneAffectedEndpointsProps {
-  insight: SpanNPlusOneInsight | undefined | null;
+  insight: SpanNPlusOneInsight | null;
 }
