+++ src/components/Insights/tickets/common/useCodeLocations.ts
@@ -1,0 +1,45 @@
+import { useEffect, useState } from "react";
+import { dispatcher } from "../../../../dispatcher";
+import { SpanInfo } from "../../../../types";
+import { actions } from "../../actions";
+import { CodeLocationsData } from "../types";
+
+export const useCodeLocations = (spanInfo: SpanInfo | null) => {
+  const [isLoading, setIsLoading] = useState(false);
+  const [codeLocations, setCodeLocations] = useState<string[]>([]);
+
+  useEffect(() => {
+    const handleCodeLocationsData = (data: unknown) => {
+      const codeLocationsData = data as CodeLocationsData;
+      setCodeLocations(codeLocationsData.codeLocations);
+      setIsLoading(false);
+    };
+
+    dispatcher.addActionListener(
+      actions.SET_CODE_LOCATIONS,
+      handleCodeLocationsData
+    );
+
+    return () => {
+      dispatcher.removeActionListener(
+        actions.SET_CODE_LOCATIONS,
+        handleCodeLocationsData
+      );
+    };
+  }, []);
+
+  useEffect(() => {
+    const spanCodeObjectId = spanInfo?.spanCodeObjectId;
+    const methodCodeObjectId = spanInfo?.methodCodeObjectId || undefined;
+
+    setIsLoading(true);
+    window.sendMessageToDigma({
+      action: actions.GET_CODE_LOCATIONS,
+      payload: {
+        spanCodeObjectId,
+        methodCodeObjectId
+      }
+    });
+  }, [spanInfo]);
+  return { isLoading, codeLocations };
+};
